add countMem and repair alais initial

+++ graph.h
@@ -11,7 +11,7 @@ struct Graph
     bool isDirected, isWeighted;
     Node *vertex;
     double stdValue, ratio;
-    unsigned int hit, miss;
+    size_t hit, miss;
 
     //根据参数设置属性并申请顶点集空间
     Graph(char** args);
